refactor: make activity card into a dialog when launching / editting
