+++ lib/pangea/common/widgets/full_width_dialog.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'package:fluffychat/config/themes.dart';
@@ -18,7 +17,7 @@ class FullWidthDialog extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final content = ConstrainedBox(
-      constraints: kIsWeb
+      constraints: FluffyThemes.isColumnMode(context)
           ? BoxConstraints(
               maxWidth: maxWidth,
               maxHeight: maxHeight,
